Extracted parent Contact class from Phone, Email and Address classes
